chore: Separate BasicReactionPickerSnapshotTests into separate file
