+++ wire-ios/Wire-iOS Tests/ReferenceImages/BasicReactionPickerSnapshotTests/test_BasicReactionPicker.1.png
