+++ wire-ios/Wire-iOS Tests/ReferenceImages/BasicReactionPickerSnapshotTests/test_BasicReactionPicker_withSelectedReaction.1.png
